[info] splits InfoSingle/Expression out from InfoBool.h
